monitoring_social/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.6
python-dateutil==2.8.2
pytz==2023.2
redis==4.5.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
vk-api==11.9.9
